anti_ageing_review/requirements.txt
pandas>=2.0
numpy>=1.24
requests>=2.31
pyyaml>=6.0
matplotlib>=3.8
seaborn>=0.13
networkx>=3.0
scikit-learn>=1.4
pytest>=8.0
